fix: changed custom dollar_rounded format key and numFmtId to 188
## Description

`dollar_2` and `dollar_rounded` defined 2 different `formatCode`, although they have the same `key` and `numFmtId` set to 183.

This cause an issue when using both formats in a sheet, where only 1 is applied because the `183` id is used for reference in the spreadsheet context.

The fix is to set `dollar_rounded` id to something else: 188 (last used key value was 187).

Also, ordered the object by key value for better readability.